+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 0.12.3 - 2020-08-22
 
 ### Changed